+++ NEWS.md
@@ -1,1 +1,10 @@
-## CHANGES IN `legisTaiwan` VERSION 1.0.1
+
+## CHANGES IN `legisTaiwan` VERSION 0.1.1
+- `get_executive_response()`, `get_bills_2()`, `get_debates()` and `get_speech_video()` are added.
+
+- lack of documentation 
+  
+
+## CHANGES IN `legisTaiwan` VERSION 0.1.0
+The package is created and had `get_meetings()`, `get_bills()`, `get_legislators()`, `get_parlquestions()`
+